Move network related collectors to network collector
