+++ pkg/gohai/collector.go
@@ -4,10 +4,8 @@ import (
 	"github.com/Datadog/verity/cpu"
 	"github.com/Datadog/verity/hostname"
 	"github.com/Datadog/verity/filesystem"
-	"github.com/Datadog/verity/ipaddress"
-	"github.com/Datadog/verity/ipv6address"
-	"github.com/Datadog/verity/macaddress"
 	"github.com/Datadog/verity/memory"
+	"github.com/Datadog/verity/network"
 	"log"
 )
 
@@ -20,10 +18,8 @@ var collectors = []Collector{
 	&cpu.Cpu{},
 	&hostname.Hostname{},
 	&filesystem.FileSystem{},
-	&ipaddress.IpAddress{},
-	&ipv6address.Ipv6Address{},
-	&macaddress.MacAddress{},
 	&memory.Memory{},
+	&network.Network{},
 }
 
 func Collect() (result map[string]interface{}, err error) {
