--- pkg/gohai/ipaddress/ipaddress.go
@@ -1,59 +1,0 @@
-package ipaddress
-
-import (
-	"errors"
-	"net"
-)
-
-type IpAddress struct{}
-
-const name = "ipaddress"
-
-func (self *IpAddress) Name() string {
-	return name
-}
-
-func (self *IpAddress) Collect() (result interface{}, err error) {
-	ipaddress, err := externalIpAddress()
-	result = ipaddress
-
-	return
-}
-
-func externalIpAddress() (string, error) {
-	ifaces, err := net.Interfaces()
-
-	if err != nil {
-		return "", err
-	}
-
-	for _, iface := range ifaces {
-		if iface.Flags&net.FlagUp == 0 || iface.Flags&net.FlagLoopback != 0 {
-			// interface down or loopback interface
-			continue
-		}
-		addrs, err := iface.Addrs()
-		if err != nil {
-			return "", err
-		}
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-			if ip == nil || ip.IsLoopback() {
-				continue
-			}
-			ip = ip.To4()
-			if ip == nil {
-				// not an ipv4 address
-				continue
-			}
-			return ip.String(), nil
-		}
-	}
-	return "", errors.New("not connected to the network")
-}
