--- pkg/gohai/macaddress/macaddress.go
@@ -1,54 +1,0 @@
-package macaddress
-
-import (
-	"errors"
-	"net"
-)
-
-type MacAddress struct{}
-
-const name = "macaddress"
-
-func (self *MacAddress) Name() string {
-	return name
-}
-
-func (self *MacAddress) Collect() (result interface{}, err error) {
-	macaddress, err := macAddress()
-	result = macaddress
-
-	return
-}
-
-func macAddress() (string, error) {
-	ifaces, err := net.Interfaces()
-
-	if err != nil {
-		return "", err
-	}
-
-	for _, iface := range ifaces {
-		if iface.Flags&net.FlagUp == 0 || iface.Flags&net.FlagLoopback != 0 {
-			// interface down or loopback interface
-			continue
-		}
-		addrs, err := iface.Addrs()
-		if err != nil {
-			return "", err
-		}
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-			if ip == nil || ip.IsLoopback() || ip.To4() == nil {
-				continue
-			}
-			return iface.HardwareAddr.String(), nil
-		}
-	}
-	return "", errors.New("not connected to the network")
-}
