+++ pkg/gohai/network/network.go
@@ -1,6 +1,7 @@
 package network
 
 import (
+	"errors"
 	"net"
 )
 
@@ -18,6 +19,151 @@ func (self *Network) Collect() (result interface{}, err error) {
 }
 
 func getNetworkInfo() (networkInfo map[string]interface{}, err error) {
+	networkInfo = make(map[string]interface{})
+
+	macaddress, err := macAddress()
+	if err != nil {
+		return networkInfo, err
+	}
+	networkInfo["macaddress"] = macaddress
+
+	ipAddress, err := externalIpAddress()
+	if err != nil {
+		return networkInfo, err
+	}
+	networkInfo["ipaddress"] = ipAddress
+
+	ipAddressV6, err := externalIpv6Address()
+	if err != nil {
+		return networkInfo, err
+	}
+	networkInfo["ipaddressv6"] = ipAddressV6
+
+	return
+}
+
+
+
+type Ipv6Address struct{}
+
+func externalIpv6Address() (string, error) {
+	ifaces, err := net.Interfaces()
+
+	if err != nil {
+		return "", err
+	}
+
+	for _, iface := range ifaces {
+		if iface.Flags&net.FlagUp == 0 || iface.Flags&net.FlagLoopback != 0 {
+			// interface down or loopback interface
+			continue
+		}
+		addrs, err := iface.Addrs()
+		if err != nil {
+			return "", err
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			if ip == nil || ip.IsLoopback() {
+				continue
+			}
+			if ip.To4() != nil {
+				// ipv4 address
+				continue
+			}
+			return ip.String(), nil
+		}
+	}
+	return "", errors.New("not connected to the network")
+}
+
+
+type IpAddress struct{}
+
+func externalIpAddress() (string, error) {
+	ifaces, err := net.Interfaces()
+
+	if err != nil {
+		return "", err
+	}
+
+	for _, iface := range ifaces {
+		if iface.Flags&net.FlagUp == 0 || iface.Flags&net.FlagLoopback != 0 {
+			// interface down or loopback interface
+			continue
+		}
+		addrs, err := iface.Addrs()
+		if err != nil {
+			return "", err
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			if ip == nil || ip.IsLoopback() {
+				continue
+			}
+			ip = ip.To4()
+			if ip == nil {
+				// not an ipv4 address
+				continue
+			}
+			return ip.String(), nil
+		}
+	}
+	return "", errors.New("not connected to the network")
+}
+
+
+type MacAddress struct{}
+
+func macAddress() (string, error) {
+	ifaces, err := net.Interfaces()
+
+	if err != nil {
+		return "", err
+	}
+
+	for _, iface := range ifaces {
+		if iface.Flags&net.FlagUp == 0 || iface.Flags&net.FlagLoopback != 0 {
+			// interface down or loopback interface
+			continue
+		}
+		addrs, err := iface.Addrs()
+		if err != nil {
+			return "", err
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			if ip == nil || ip.IsLoopback() || ip.To4() == nil {
+				continue
+			}
+			return iface.HardwareAddr.String(), nil
+		}
+	}
+	return "", errors.New("not connected to the network")
+}
+
+
+/* Not currently used */
+
+func getDetailedNetworkInfo() (networkInfo map[string]interface{}, err error) {
 	ifaces, err := net.Interfaces()
 
 	if err != nil {
